requirements.txt
python-magic==0.4.24
Pillow==8.4.0
